Create Vowel_Formants
This script is used to extract formant values from labels containing specific values (e.g. Here, I want to extract from labels that are marked with u, i, y).
This script will loop through all the files in a folder, and will extract F1/F2/F3 values from timepoints of: (1) 50%; or (2) 20% and 35% and 50% and 65% and 80% from intervals that contain u, i, y.